FindAvailableDogsBySize UseCase Done

+++ src/DogShelter.Domain/Entities/BreedEntity/Breed.cs
@@ -5,14 +5,14 @@ namespace DogShelter.Domain.Entities.BreedEntity;
 
 public class Breed : BaseEntity
 {
-    public int    ApiId                 { get; set; }
-    public string Name                  { get; set; } = null!;
-    public string BredFor               { get; set; } = null!;
-    public string BreedGroup            { get; set; } = null!;
-    public string LifeSpan              { get; set; } = null!;
-    public string Temperament           { get; set; } = null!;
-    public int    HeightAverageMetric   { get; set; }
-    public int    HeightAverageImperial { get; set; }
+    public int     ApiId                 { get; set; }
+    public string  Name                  { get; set; } = null!;
+    public string  BredFor               { get; set; } = null!;
+    public string  BreedGroup            { get; set; } = null!;
+    public string  LifeSpan              { get; set; } = null!;
+    public string  Temperament           { get; set; } = null!;
+    public int     HeightAverageMetric   { get; set; }
+    public decimal HeightAverageImperial { get; set; }
 
     public List<Dog> Dogs { get; set; }
 }
